pyproject.toml
[tool.poetry]
name = "todo app"
version = "0.1.0"
description = "Todo App using Python and Django."
authors = ["RajeshKumar <https://github.com/Rajeshkumar-14>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
django-ajax-datatable = "^4.4.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
